Split publickey resolving to a dedicated class for better testing
